Create README.md
CALL / WHATSAPP +27710515041 Safe abortion pills for sale in Bulawayo Gweru Kwekwe Harare
CALL / WHATSAPP +27710515041 MISSED YOUR MONTHLY PERIODS? DON'T STRESS, DR DONALD WOMEN'S ABORTION CLINIC THOHOYANDOU We have a 24 hour phone service in case of any concerns. About 90% of women feel relieved following an elective abortion, and engage positively with their lives Women's Abortion Clinic offers same day services that are safe and pain free. We use approved pills and we clean the womb so that no side effects are present, We deliver pills nationwide for those who wish to undergo the process in the comfort of their homes. We offer Abortion services including "Abortions 1 to 28 weeks and beyond, Abortions by Pill and The Morning After Pill." we also offer pregnancy testing, and a full range of other services. Beyond other abortion clinics, we specialize in the treatment of pelvic pain, painful fibroids and heavy periods. We are proud of our reputation for being a medical office that treats each patient with distinctive care and respect. The decision to terminate a pregnancy is not an easy one to make. But, if you have already decided to have an abortion, be assured that you are not alone. We understand what it's like for you if you have an unplanned pregnancy. You may be confused and uncertain where to go for help. DELIVERIES CAN BE MADE AT YOUR CONVENIENCE THROUGH ARRANGEMENTS, DO NOT HESITATE TO CALL: +27710515041 We respect your right to privacy and all consultations are confidential. WE DO TERMINATION FROM ONE WEEK TO TWENTY EIGHT WEEKS. DR DONALD WOMEN'S ABORTION CLINIC IN Nyanga, , Nyazura, Odzi, Birchenough Bridge